Add dateOfBirth and age to child sensitive information page
- Also fix the bloated layout to match Figma
- Add dateOfBirth to ChildSensitiveInformation
- Remove old sensitive info endpoint
- Move child sensitive info features to their own package "sensitive"
- Remove redundant ResponseEntity usage
