+++ frontend/src/employee-mobile-frontend/components/attendances/child-info/ChildSensitiveInfoPage.tsx
@@ -8,7 +8,7 @@ import { faArrowLeft } from 'lib-icons'
 import React, { useContext, useMemo } from 'react'
 import { useHistory, useParams } from 'react-router-dom'
 import styled from 'styled-components'
-import { getChildSensitiveInformation } from '../../../api/attendances'
+import { getChildSensitiveInformation } from '../../../api/sensitive'
 import { ChildAttendanceContext } from '../../../state/child-attendance'
 import { useTranslation } from '../../../state/i18n'
 import { renderPinRequiringResult } from '../../auth/renderPinRequiringResult'
